los que devuelve el metodo get valo,  int por long

+++ SUPROPIEDADRAIZ/src/suPropiedadRaiz/contratos/CompraVenta.java
@@ -10,7 +10,7 @@ public class CompraVenta {
 	Apartamento apartamento;
 	Cliente clienteOferente, clienteComprador;
 	
-	public int getValor() {
+	public long getValor() {
 		return valor;
 	}
